+++ apps/fluent-tester/src/E2E/Separator/specs/Separator.spec.win.ts
@@ -1,4 +1,4 @@
-import NavigateAppPage from '../../common/NavigateAppPage';
+import NavigateAppPage from '../../common/NavigateAppPage.win';
 import SeparatorPageObject from '../pages/SeparatorPageObject.win';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';