+++ apps/fluent-tester/src/E2E/Tabs/specs/Tabs.spec.win.ts
@@ -1,4 +1,4 @@
-import NavigateAppPage from '../../common/NavigateAppPage';
+import NavigateAppPage from '../../common/NavigateAppPage.win';
 import TabsPageObject from '../pages/TabsPageObject.win';
 import { TAB_A11Y_ROLE, BOOT_APP_TIMEOUT, PAGE_TIMEOUT, TABITEM_A11Y_ROLE } from '../../common/consts';