+++ apps/fluent-tester/src/E2E/Theme/pages/ThemePageObject.win.ts
@@ -1,5 +1,5 @@
 import { THEME_TESTPAGE, THEME_TEST_COMPONENT, HOMEPAGE_THEME_BUTTON } from '../../../FluentTester/TestComponents/Theme/consts';
-import { BasePage, By } from '../../common/BasePage';
+import { BasePage, By } from '../../common/BasePage.win';
 
 class ThemePageObject extends BasePage {
   /*****************************************/