+++ packages/components/ContextualMenu/src/ContextualMenuItem.tsx
@@ -1,6 +1,6 @@
 /** @jsx withSlots */
 import * as React from 'react';
-import { View } from 'react-native';
+import { Platform, View } from 'react-native';
 import {
   ContextualMenuItemSlotProps,
   ContextualMenuItemState,
@@ -15,7 +15,7 @@ import { Text } from '@fluentui-react-native/text';
 import { settings } from './ContextualMenuItem.settings';
 import { backgroundColorTokens, borderTokens, textTokens, foregroundColorTokens, getPaletteFromTheme } from '@fluentui-react-native/tokens';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
-import { useAsPressable, useKeyCallback, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
+import { useAsPressable, useKeyUpProps, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
 import { CMContext } from './ContextualMenu';
 import { Icon } from '@fluentui-react-native/icon';
 import { createIconProps } from '@fluentui-react-native/interactive-hooks';
@@ -63,7 +63,7 @@ export const ContextualMenuItem = compose<ContextualMenuItemType>({
 
     const pressable = useAsPressable({ ...rest, onPress: onItemClick, onHoverIn: onItemHoverIn });
 
-    const onKeyUp = useKeyCallback(onItemClick, ' ', 'Enter');
+    const onKeyUpProps = useKeyUpProps(onItemClick, ' ', 'Enter');
 
     // set up state
     const state: ContextualMenuItemState = {
@@ -103,10 +103,25 @@ export const ContextualMenuItem = compose<ContextualMenuItemType>({
       root: {
         ...pressable.props,
         ref: cmRef,
-        onKeyUp: onKeyUp,
-        onMouseEnter: onMouseEnter,
-        onMouseLeave: onMouseLeave,
+        ...onKeyUpProps,
+        onMouseEnter,
+        onMouseLeave,
+        accessible: true,
         accessibilityLabel: accessibilityLabel,
+        accessibilityRole: 'menuitem',
+        accessibilityState: { disabled: state.disabled, selected: state.selected },
+        accessibilityValue: { text: itemKey },
+        focusable: Platform.select({
+          /**
+           * GH #1208: On macOS, disabled NSMenuItems are not focusable unless VoiceOver is enabled.
+           * To mimic the non VoiceOver-enabled functionality, let's set focusable to !disabled for now.
+           * As a followup, we could query AccessibilityInfo.isScreenReaderEnabled, and pass that info to
+           * CMContext so that the event handler is only handled once per ContextualMenu.
+           */
+          macos: !disabled,
+          // Keep win32 behavior as is
+          default: true,
+        }),
         ...rest,
       },
       content: { children: text },