+++ packages/components/ContextualMenu/src/SubmenuItem.tsx
@@ -1,6 +1,6 @@
 /** @jsx withSlots */
 import * as React from 'react';
-import { View } from 'react-native';
+import { Platform, View } from 'react-native';
 import {
   SubmenuItemSlotProps,
   SubmenuItemState,
@@ -15,7 +15,7 @@ import { Text } from '@fluentui-react-native/text';
 import { settings } from './SubmenuItem.settings';
 import { backgroundColorTokens, borderTokens, textTokens, foregroundColorTokens, getPaletteFromTheme } from '@fluentui-react-native/tokens';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
-import { useAsPressable, useKeyCallback, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
+import { useAsPressable, useKeyUpProps, useViewCommandFocus } from '@fluentui-react-native/interactive-hooks';
 import { SvgXml } from 'react-native-svg';
 import { CMContext } from './ContextualMenu';
 import { Icon } from '@fluentui-react-native/icon';
@@ -83,7 +83,7 @@ export const SubmenuItem = compose<SubmenuItemType>({
     /*
      * SubmenuItem launches the submenu onMouseEnter event. For keyboarding, submenu should be launched with Spacebar, Enter, or right arrow.
      */
-    const onKeyUp = useKeyCallback(onMouseEnter, ' ', 'Enter', 'ArrowRight');
+    const onKeyUpProps = useKeyUpProps(onMouseEnter, ' ', 'Enter', 'ArrowRight');
 
     // grab the styling information, referencing the state as well as the props
     const styleProps = useStyling(userProps, (override: string) => state[override] || userProps[override]);
@@ -92,9 +92,24 @@ export const SubmenuItem = compose<SubmenuItemType>({
       root: {
         ...pressable.props,
         ref: cmRef,
-        onKeyUp,
+        ...onKeyUpProps,
         onMouseEnter,
-        accessibilityLabel,
+        accessible: true,
+        accessibilityLabel: accessibilityLabel,
+        accessibilityRole: 'menuitem',
+        accessibilityState: { disabled: state.disabled, selected: state.selected },
+        accessibilityValue: { text: itemKey },
+        focusable: Platform.select({
+          /**
+           * GH #1208: On macOS, disabled NSMenuItems are not focusable unless VoiceOver is enabled.
+           * To mimic the non VoiceOver-enabled functionality, let's set focusable to !disabled for now.
+           * As a followup, we could query AccessibilityInfo.isScreenReaderEnabled, and pass that info to
+           * CMContext so that the event handler is only handled once per ContextualMenu.
+           */
+          macos: !disabled,
+          // Keep win32 behavior as is
+          default: true,
+        }),
         ...rest,
       },
       content: { children: text },
@@ -107,38 +122,38 @@ export const SubmenuItem = compose<SubmenuItemType>({
   settings,
   render: (Slots: ISlots<SubmenuItemSlotProps>, renderData: SubmenuItemRenderData, ...children: React.ReactNode[]) => {
     const xml = `
-    <svg width="12" height="12" viewBox="0 0 2048 2048" >
+    <svg width="12" height="12" viewBox="0 0 2048 2048">
       <path class='OfficeIconColors_HighContrast' fill='currentColor' d='M 743 1767 l -121 -121 l 708 -707 l -708 -708 l 121 -121 l 828 829 z' />
-      <path class='OfficeIconColors_m22' fill='currentColor' d='M 743 1767 l -121 -121 l 708 -707 l -708 -708 l 121 -121 l 828 829 z' />
+      <path class='OfficeIconColors_m22'          fill='currentColor' d='M 743 1767 l -121 -121 l 708 -707 l -708 -708 l 121 -121 l 828 829 z' />
     </svg>`;
     // We shouldn't have to specify the source prop on Slots.icon, here, but we need another drop from @uifabricshared
     return (
       <Slots.root>
-        <Slots.leftstack>
+        <Slots.startstack>
           {renderData!.state.icon && <Slots.icon />}
           {renderData!.state.content && <Slots.content />}
           {children}
-        </Slots.leftstack>
-        <Slots.rightstack>
+        </Slots.startstack>
+        <Slots.endstack>
           <Slots.chevron xml={xml} />
-        </Slots.rightstack>
+        </Slots.endstack>
       </Slots.root>
     );
   },
   slots: {
     root: View,
-    leftstack: { slotType: View },
-    icon: { slotType: Icon as React.ComponentType },
+    startstack: View,
+    icon: Icon as React.ComponentType,
     content: Text,
-    rightstack: { slotType: View },
+    endstack: View,
     chevron: SvgXml,
   },
   styles: {
     root: [backgroundColorTokens, borderTokens],
-    leftstack: [],
+    startstack: [],
     icon: [{ source: 'iconColor', lookup: getPaletteFromTheme, target: 'color' }],
     content: [textTokens, foregroundColorTokens],
-    rightstack: [],
+    endstack: [],
     chevron: [{ source: 'chevronColor', lookup: getPaletteFromTheme, target: 'color' }],
   },
 });