+++ packages/experimental/Button/src/Button.styling.ts
@@ -1,4 +1,4 @@
-import { buttonName, ButtonCoreTokens, ButtonTokens, ButtonSlotProps, ButtonProps, ButtonSize } from './Button.types';
+import { buttonName, ButtonCoreTokens, ButtonTokens, ButtonSlotProps, ButtonPropsWithInnerRef, ButtonSize } from './Button.types';
 import { Theme, UseStylingOptions, buildProps } from '@fluentui-react-native/framework';
 import { borderStyles, layoutStyles, fontStyles, shadowStyles } from '@fluentui-react-native/tokens';
 import { defaultButtonTokens } from './ButtonTokens';
@@ -13,7 +13,6 @@ export const buttonStates: (keyof ButtonTokens)[] = [
   'primary',
   'subtle',
   'hovered',
-  'focused',
   'pressed',
   'disabled',
   'small',
@@ -24,9 +23,10 @@ export const buttonStates: (keyof ButtonTokens)[] = [
   'rounded',
   'circular',
   'square',
+  'focused',
 ];
 
-export const stylingSettings: UseStylingOptions<ButtonProps, ButtonSlotProps, ButtonTokens> = {
+export const stylingSettings: UseStylingOptions<ButtonPropsWithInnerRef, ButtonSlotProps, ButtonTokens> = {
   tokens: [defaultButtonTokens, defaultButtonColorTokens, buttonName],
   states: buttonStates,
   slotProps: {
@@ -48,14 +48,22 @@ export const stylingSettings: UseStylingOptions<ButtonProps, ButtonSlotProps, Bu
       ['backgroundColor', 'width', ...borderStyles.keys, ...layoutStyles.keys, ...shadowStyles.keys],
     ),
     content: buildProps(
-      (tokens: ButtonTokens, theme: Theme) => ({
-        style: {
-          color: tokens.color,
-          ...getTextMarginAdjustment(),
-          ...(tokens.spacingIconContent && { marginLeft: tokens.spacingIconContent }),
-          ...fontStyles.from(tokens, theme),
-        },
-      }),
+      (tokens: ButtonTokens, theme: Theme) => {
+        const spacingIconContent = tokens.spacingIconContent
+          ? {
+              marginLeft: tokens.spacingIconContent,
+              marginRight: tokens.spacingIconContent,
+            }
+          : {};
+        return {
+          style: {
+            color: tokens.color,
+            ...getTextMarginAdjustment(),
+            ...spacingIconContent,
+            ...fontStyles.from(tokens, theme),
+          },
+        };
+      },
       ['color', 'spacingIconContent', ...fontStyles.keys],
     ),
     icon: buildProps(