+++ packages/experimental/Button/src/Button.tsx
@@ -1,13 +1,14 @@
 /** @jsx withSlots */
 import * as React from 'react';
 import { View } from 'react-native';
-import { buttonName, ButtonType, ButtonProps } from './Button.types';
+import { ActivityIndicator } from '@fluentui-react-native/experimental-activity-indicator';
+import { buttonName, ButtonType, ButtonProps, ButtonPropsWithInnerRef } from './Button.types';
 import { Text } from '@fluentui-react-native/experimental-text';
 import { stylingSettings, getDefaultSize } from './Button.styling';
 import { compose, mergeProps, withSlots, UseSlots } from '@fluentui-react-native/framework';
 import { useButton } from './useButton';
 import { Icon } from '@fluentui-react-native/icon';
-import { createIconProps, IPressableState } from '@fluentui-react-native/interactive-hooks';
+import { createIconProps, IFocusable, IPressableState } from '@fluentui-react-native/interactive-hooks';
 
 /**
  * A function which determines if a set of styles should be applied to the compoent given the current state and props of the button.
@@ -17,7 +18,7 @@ import { createIconProps, IPressableState } from '@fluentui-react-native/interac
  * @param userProps The props that were passed into the button
  * @returns Whether the styles that are assigned to the layer should be applied to the button
  */
-export const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonProps): boolean => {
+export const buttonLookup = (layer: string, state: IPressableState, userProps: ButtonPropsWithInnerRef): boolean => {
   return (
     state[layer] ||
     userProps[layer] ||
@@ -26,12 +27,12 @@ export const buttonLookup = (layer: string, state: IPressableState, userProps: B
     (!userProps['size'] && layer === getDefaultSize()) ||
     layer === userProps['shape'] ||
     (!userProps['shape'] && layer === 'rounded') ||
-    (layer === 'hasContent' && userProps.content) ||
-    (layer === 'hasIcon' && userProps.icon)
+    (layer === 'hasContent' && !userProps.iconOnly) ||
+    (layer === 'hasIcon' && (userProps.icon || userProps.loading))
   );
 };
 
-export const Button = compose<ButtonType>({
+const ButtonComposed = compose<ButtonType>({
   displayName: buttonName,
   ...stylingSettings,
   slots: {
@@ -39,25 +40,29 @@ export const Button = compose<ButtonType>({
     icon: Icon,
     content: Text,
   },
-  render: (userProps: ButtonProps, useSlots: UseSlots<ButtonType>) => {
+  render: (userProps: ButtonPropsWithInnerRef, useSlots: UseSlots<ButtonType>) => {
     const button = useButton(userProps);
     const iconProps = createIconProps(userProps.icon);
     // grab the styled slots
     const Slots = useSlots(userProps, (layer) => buttonLookup(layer, button.state, userProps));
 
     // now return the handler for finishing render
-    return (final: ButtonProps, ...children: React.ReactNode[]) => {
-      const { icon, iconPosition, content, ...mergedProps } = mergeProps(button.props, final);
+    return (final: ButtonPropsWithInnerRef, ...children: React.ReactNode[]) => {
+      const { icon, iconPosition, loading, ...mergedProps } = mergeProps(button.props, final);
       return (
         <Slots.root {...mergedProps}>
           {icon && iconPosition === 'before' && <Slots.icon {...iconProps} />}
-          {content && <Slots.content key="content">{content}</Slots.content>}
-          {children}
+          {loading && <ActivityIndicator />}
+          {React.Children.map(children, (child) =>
+            typeof child === 'string' ? <Slots.content key="content">{child}</Slots.content> : child,
+          )}
           {icon && iconPosition === 'after' && <Slots.icon {...iconProps} />}
         </Slots.root>
       );
     };
   },
 });
+
+export const Button = React.forwardRef<IFocusable, ButtonProps>((props, ref) => <ButtonComposed {...props} innerRef={ref} />);
 
 export default Button;