+++ packages/experimental/Button/src/CompoundButton/CompoundButton.types.ts
@@ -1,6 +1,6 @@
 import { ViewProps, ColorValue } from 'react-native';
 import { TextProps } from '@fluentui-react-native/experimental-text';
-import { ButtonSlotProps, ButtonTokens, ButtonProps } from '../Button.types';
+import { ButtonSlotProps, ButtonTokens, ButtonPropsWithInnerRef } from '../Button.types';
 import { FontTokens } from '@fluentui-react-native/tokens';
 
 export const compoundButtonName = 'CompoundButton';
@@ -24,12 +24,14 @@ export interface CompoundButtonTokens extends ButtonTokens {
   hasContent?: CompoundButtonTokens;
 }
 
-export interface CompoundButtonProps extends ButtonProps {
+export interface CompoundButtonPropsWithInnerRef extends ButtonPropsWithInnerRef {
   /**
    * Second line of text that describes the action this button takes.
    */
   secondaryContent?: string;
 }
+
+export type CompoundButtonProps = Omit<CompoundButtonPropsWithInnerRef, 'innerRef'>;
 
 export interface CompoundButtonSlotProps extends ButtonSlotProps {
   contentContainer: ViewProps;
@@ -37,7 +39,7 @@ export interface CompoundButtonSlotProps extends ButtonSlotProps {
 }
 
 export interface CompoundButtonType {
-  props: CompoundButtonProps;
+  props: CompoundButtonPropsWithInnerRef;
   tokens: CompoundButtonTokens;
   slotProps: CompoundButtonSlotProps;
 }