+++ packages/experimental/Button/src/FAB/FAB.types.ts
@@ -1,9 +1,9 @@
-import { ButtonSlotProps, ButtonCoreTokens, ButtonCoreProps } from '../Button.types';
+import { ButtonSlotProps, ButtonCoreTokens, ButtonCorePropsWithInnerRef } from '../Button.types';
 
 export const fabName = 'FAB';
 
 export interface FABType {
-  props: ButtonCoreProps;
+  props: ButtonCorePropsWithInnerRef;
   tokens: ButtonCoreTokens;
   slotProps: ButtonSlotProps;
 }