+++ packages/theming/theme-tokens/src/tokens-global.android.ts
@@ -1,3 +1,3 @@
-import globalTokens from './global-android/reactnative/tokens-global.json';
+import globalTokens from './generated/global-android/reactnative/tokens-global.json';
 
 export default globalTokens;