+++ packages/theming/theme-tokens/src/tokens-global.ts
@@ -1,3 +1,3 @@
-import globalTokens from './global/reactnative/tokens-global.json';
+import globalTokens from './generated/global/reactnative/tokens-global.json';
 
 export default globalTokens;